[IMP] Improve pertinent change detection between commits
The changes we make on a branch are not 100% of the time reflected in
the bundle `o_spreadsheet.js`(i.e. changes in the src folder). In those
cases, while using `update` we create useless PRs that will only change
the last hash and are often just closed.

This commit improves the detection of changes to skip versions which do
not contain relevant changes in the bundle.