[IMP] Followup of move to Odoo
Even though we now merge on Odoo, we still need to create an enterprise
branch as some features in Enterprise (e.g. the edition, the templates)
depend on the update of the library in Community and we need to run
their tests as well before validating the lib update in Community.