[IMP] update:  fallback on latest release commit
the script currently skips a version if its last commit is not a [REL]
commit, this is bothersome because it forces you to run release then
update without anyone else merging anything.

This commit removes that limitation and goes back to the lastest release
commit in o-spreadsheet to update odoo community/enterprise.