[IMP] update: Don't create empty branches
Runbot does not like that we create empty branches for the sake of it.
Our goal was to have all the tests run automatically to ensure we don't
break anything when updating the library but the existence of branches
which have common commits with the stable branches is apparently
problematic.

This revision adds the creation of a ghost commit in the enterprise
branch to address that issue.