[IMP] release: allow key users to self validate releases
The release script only creates an automated commit that will ensure the
release on GH and publication on NPM. The integrity of the changes has
been reviewed in prior pull requests. As such, the release commmit can
be trusted and self-reviewed by their creator - which will only be done
by a manager of the spreadsheet project.