[FIX] config: fix version alignment
The version alignment was exectuted on the current working directory,
which means that if called `update` or  `release` from another directory
than the  root of this project, you'd be checking irrelevant hash and
have false positive/negative results depending on your luck.

Moreover, this would only check if you were up-to-date with your own
remote branch and not the master, which again went against the goal of
this check.