[IMP] transpile ahead-of-time
In odoo, when asset bundles are generated, file undergo a transpilation
process to convert the imports and exports to `odoo.define`.

The `o_spreadsheet.js` being very large, running this process is slow.

This commit adds a step after copying the lib file to odoo to transpile
it ahead of time.

Time to generate the spreadsheet asset bundle:

before: ~1.1s
after: ~0.55s